Restructures tests into photometric and spectroscopic sets
