+++ tests/test_snls_balland09.py
@@ -1,34 +1,22 @@
 #!/usr/bin/env python3
 # -*- coding: UTF-8 -*-
 
-"""Tests for the ``snls.balland09`` module."""
+"""Tests for the ``snls`` module"""
+
+from unittest import TestCase, skip
 
 from sndata import snls
-from .base_tests import DataParsingTestBase
+from . import template_tests
 
 
-class DataParsing(DataParsingTestBase):
-    """Tests for the snls.balland09 module"""
+class Balland09(TestCase, template_tests.SpectroscopicDataParsing):
+    """Generic tests for a given survey"""
 
     @classmethod
     def setUpClass(cls):
         cls.test_class = snls.Balland09()
         cls.test_class.download_module_data()
 
-    def test_bad_object_id_err(self):
-        self._test_bad_object_id_err()
-
-    def test_ids_are_sorted(self):
-        self._test_ids_are_sorted()
-
-    def test_no_empty_data_tables(self):
-        self._test_no_empty_data_tables(10)
-
-    def test_paper_tables_are_parsed(self):
-        self._test_paper_tables_are_parsed()
-
-    def test_unique_ids(self):
-        self._test_unique_ids()
-
-    def test_cache_not_mutated(self):
-        self._test_cache_not_mutated()
+    skip('Balland09 does not have dates in the data tables')
+    def test_jd_time_format(self):
+        pass
